Trab2Day2API/pythonUnimar/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-dotenv==1.1.0
openai==1.84.0
anthropic==0.52.2
pydantic==2.11.5
httpx==0.28.1
typing_extensions==4.14.0
